RivelyPY/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
openai==1.50.2
pydantic==2.9.2
pydantic_core==2.23.4
requests==2.32.3
scrapingbee==2.0.1
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
